Deploy Submission

+++ backend/image/src/upload/upload.service.ts
@@ -7,9 +7,9 @@ import {
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { createHash } from 'crypto';
-import { existsSync, rmSync, lstatSync } from 'fs';
-import { copyFile, rm, readFile, writeFile } from 'fs/promises';
-import {join} from 'path';
+import { constants, copyFileSync, existsSync, lstatSync, rmSync } from 'fs';
+import { copyFile, readFile, writeFile } from 'fs/promises';
+import { join } from 'path';
 import * as sharp from 'sharp';
 import { InternalCommunicationService } from '../internal-communication/internal-communication.service';
 import {
@@ -25,6 +25,7 @@ import { ImageUploadResponse } from './responses/upload-image.response';
 export class UploadService {
   private storePath: string;
   private cachePath: string;
+  private tmpPath: string;
   private readonly logger = new Logger(UploadService.name);
   constructor(
     private readonly configService: ConfigService,
@@ -40,7 +41,13 @@ export class UploadService {
       'cache',
     );
     validateOrCreateDirectory(this.cachePath);
+    this.tmpPath = pathBuilder(
+      this.configService.get<string>('IMAGE_STORE_TMP_PATH'),
+      'tmp',
+    );
+    validateOrCreateDirectory(this.tmpPath);
   }
+
   /**
    * @description Handle uploaded image including duplicate check, hashing and saving it
    */
@@ -50,47 +57,44 @@ export class UploadService {
   ): Promise<ImageUploadResponse> {
     const cachedFilePath = join(this.cachePath, image.filename);
 
+    const fileType = `${image.mimetype.split('/')[1]}`;
+
     // Crop the image before hashing => otherwise hashing would be useless
     await this.cropImage(cachedFilePath, 512, 512);
 
-    // TODO: Remove all geographic data etc. from image file
+    const hash = `${await this.hashFile(cachedFilePath)}.${fileType}`;
+    const fileTargetPath = join(this.tmpPath, hash);
 
-    const hash = await this.hashFile(cachedFilePath);
-    const fileTargetPath = join(
-      this.storePath,
-      `${hash}.${image.mimetype.split('/')[1]}`,
-    );
+    // Does file exist in permanent or temporary storage?
+    if (existsSync(fileTargetPath) || existsSync(join(this.storePath, hash))) {
+      this.logger.log(`Image ${hash} already exists`);
+      // Image is duplicate => return error along with the hash => no duplicate files
+      throw new ConflictException({
+        message:
+          'This file already seems to be uploaded (indicated by m5hash of file)',
+        path: hash,
+      } as ImageUploadConflictError);
+    }
 
     try {
-      if (existsSync(fileTargetPath)) {
-        // Image is duplicate => return error along with the hash => no duplicate files
-        throw new ConflictException({
-          message:
-            'This file already seems to be uploaded (indicated by m5hash of file)',
-          path: `${hash}.${image.mimetype.split('/')[1]}`,
-        } as ImageUploadConflictError);
-      }
       // Copy rather than move to allow for "moving" accross devices (i.e. docker volumes)
-      this.logger.debug('Promoting image from cache to disk');
+      this.logger.debug('Promoting image from cache to tmp');
       await copyFile(cachedFilePath, fileTargetPath);
-    } catch (error) {
+    } /* istanbul ignore next */ catch (error) {
       this.logger.error(error);
-      if (error instanceof HttpException) {
-        throw error;
-      }
       throw new InternalServerErrorException();
     } finally {
       // Cleanup cache
-      await rm(cachedFilePath, { force: true });
+      await rmSync(cachedFilePath, { force: true });
     }
 
     try {
       await this.internalCommunicationService.notifyAuthAboutNewImage(
         userJWT,
-        `${hash}.${image.mimetype.split('/')[1]}`,
+        hash,
       );
     } catch (error) {
-      this.logger.warn(
+      this.logger.error(
         `An upload failed because the communication to the  auth backend failed. This could indicate a crashed auth service. ${error}`,
       );
       // A file without an owner is not allowed => cleanup
@@ -102,7 +106,7 @@ export class UploadService {
     }
 
     this.logger.log('Sucesfully uploaded Image with hash: ', hash);
-    return { path: `${hash}.${image.mimetype.split('/')[1]}` };
+    return { path: hash };
   }
 
   /**
@@ -124,38 +128,97 @@ export class UploadService {
     sourcePath: string,
     wDimension: number,
     hDimension: number,
-  ) {
+  ): Promise<void> {
     this.logger.debug(`Cropping image ${sourcePath}`);
+
     if (!existsSync(sourcePath)) {
-      throw new InternalServerErrorException(
-        'Image could not be found within cache',
+      this.logger.error(
+        `Image could not be cropped as it could not be found in cache (${sourcePath})`,
       );
+      throw new InternalServerErrorException('Image cropping failed');
     }
+
     if (lstatSync(sourcePath).isDirectory()) {
-      throw new InternalServerErrorException('Cannot crop directory');
+      this.logger.error(
+        `Image cropping path points to a directory (${sourcePath})`,
+      );
+      throw new InternalServerErrorException('Image cropping failed');
     }
+
     const image = sharp(sourcePath);
     const metadata = await image.metadata();
+
     // Image is smaller or equal to the size limit
     if (metadata.width <= wDimension && metadata.height <= hDimension) {
       return;
     }
+
     // Calculate offset so the image is cropped symetrically
     const wOffset = Math.floor((metadata.width - wDimension) / 2);
     const hOffset = Math.floor((metadata.height - hDimension) / 2);
     try {
       const buffer = await image
         .extract({
-          width: Math.min(wDimension, metadata.width),
-          height: Math.min(hDimension, metadata.height),
-          left: wOffset,
-          top: hOffset,
+          width: Math.max(0, Math.min(wDimension, metadata.width)),
+          height: Math.max(0, Math.min(hDimension, metadata.height)),
+          left: Math.max(0, wOffset),
+          top: Math.max(0, hOffset),
         })
         .toBuffer();
       await writeFile(sourcePath, buffer);
-    } catch (error) {
+    } /* istanbul ignore next */ catch (error) {
       this.logger.error(error);
       throw new InternalServerErrorException('Image cropping failed');
+    }
+  }
+
+  /**
+   * @description Move image from temporary to permanent storage
+   */
+  async promoteImage(imageHash: string): Promise<void> {
+    this.logger.debug(`Promoting image ${imageHash} from tmp to permanent`);
+    const currentPath = join(this.tmpPath, imageHash);
+    const targetPath = join(this.storePath, imageHash);
+    if (!existsSync(currentPath)) {
+      this.logger.log(
+        `Image could not be promoted as it was not in temporary storage. This may indicate that the image has already been promoted or deleted. However this is an expected edge case.`,
+      );
+      return;
+    }
+    if (existsSync(targetPath)) {
+      this.logger.log('Image already exists in promoted form');
+      return;
+    }
+    this.moveFile(currentPath, targetPath);
+  }
+
+  /**
+   * @description Move image from permanent to temporary storage
+   */
+  async demoteImage(imageHash: string): Promise<void> {
+    this.logger.debug(`Demoting image ${imageHash} from permanent to tmp`);
+    const currentPath = join(this.storePath, imageHash);
+    const targetPath = join(this.tmpPath, imageHash);
+    if (!existsSync(currentPath) || existsSync(targetPath)) {
+      this.logger.log(`Image demotion failed ${imageHash}`);
+      return;
+    }
+    this.moveFile(currentPath, targetPath);
+  }
+
+  /**
+   * @description Move image (allow move accross devices)
+   */
+  private moveFile(source: string, target: string): void {
+    try {
+      // Use this instead of move to allow for "moving" accross devices (i.e. in a docker volume environment)
+      copyFileSync(source, target, constants.COPYFILE_EXCL);
+      rmSync(source);
+    } catch (error) {
+      this.logger.error(
+        `Move from ${source} to ${target} failed due to an error ${error}`,
+      );
+      throw new InternalServerErrorException('Image could not be moved');
     }
   }
 }
